fix: remove broken optimisation
Using the Viper code emitter for optimised bitfield access saves ~0.3ms
in the update loop, but breaks the arithmetic and results (somehow) in
gyro drift of approx. 0.5 deg/s. (Maybe an issue with the return
conversion?)

This commit disables the Viper emitter on the `BitField` class and fixes
the drift issues.